# Speed up `test_ingestion.py` tests by running on less data, mix up dimensions size more between tests