@jparismorgan 2024-05-15
> I think that we can also improve `test/test_object_index.py::test_object_index_ivf_flat_cloud` which is the test that can take a lot of latency especially when cloud resources are not available

Yep agree good call. I took a quick look when making this PR but didn't see an easy knob to tune, but would be awesome if we can get it down to the 300->500 second range as well.